+++ member-service/src/main/java/com/rainbowgon/memberservice/domain/client/SearchServiceClient.java
@@ -1,7 +1,7 @@
-package com.rainbowgon.memberservice.domain.bookmark.client;
+package com.rainbowgon.memberservice.domain.client;
 
-import com.rainbowgon.memberservice.domain.bookmark.client.dto.input.SearchDetailInDto;
-import com.rainbowgon.memberservice.domain.bookmark.client.dto.input.SearchSimpleInDto;
+import com.rainbowgon.memberservice.domain.client.dto.input.SearchDetailInDto;
+import com.rainbowgon.memberservice.domain.client.dto.input.SearchSimpleInDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -12,9 +12,9 @@ import java.util.Set;
 @FeignClient(name = "search-service")
 public interface SearchServiceClient {
 
-    @GetMapping("/themes/info")
+    @GetMapping("/clients/themes/info")
     List<SearchDetailInDto> getBookmarkThemeDetailInfo(@RequestBody List<String> themeIdList);
 
-    @GetMapping("/themes/")
+    @GetMapping("/clients/themes/")
     List<SearchSimpleInDto> getBookmarkThemeSimpleInfo(@RequestBody Set<String> themeIdSet);
 }
