+++ member-service/src/main/java/com/rainbowgon/memberservice/domain/client/dto/input/ReservationInDto.java
@@ -1,0 +1,20 @@
+package com.rainbowgon.memberservice.domain.client.dto.input;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ReservationInDto {
+
+    private Long reservationId; // 예약 ID
+    private Long reservationNumber; // 예약 번호
+    private String themeId; // 테마 ID
+    private LocalDate targetDate; // 예약 날짜
+    private LocalTime targetTime; // 예약 시간
+    private Integer headcount; // 인원
+}
