+++ member-service/src/main/java/com/rainbowgon/memberservice/domain/client/dto/input/SearchSimpleInDto.java
@@ -1,0 +1,16 @@
+package com.rainbowgon.memberservice.domain.client.dto.input;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class SearchSimpleInDto {
+
+    private String themeId; // 테마 ID
+    private String poster; // 테마 포스터
+    private String title; // 테마명
+    private String venue; // 지점명
+    private String location; // 지점 위치
+}
