+++ member-service/src/main/java/com/rainbowgon/memberservice/domain/client/dto/output/NotificationOutDto.java
@@ -1,4 +1,4 @@
-package com.rainbowgon.memberservice.domain.bookmark.client.dto.output;
+package com.rainbowgon.memberservice.domain.client.dto.output;
 
 import lombok.*;
